+++ CHANGELOG.md
@@ -5,7 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.7.1] - Next version
+## [0.7.x] - Next version
+
+* Get list of client role mappings for a group
+* Save client role mappings for a user/group
+* Save realm-level role mappings for a user/group
+
+## [0.7.1] - 2019-06-11
 
 Thanks to @vlad-ro:
 
@@ -18,7 +24,7 @@ Thanks to @vlad-ro:
 * Support passing rest client options for user save and search
 * Support using gem without ActiveSupport
 
-## [0.7.0] - 2019-06-11
+## [0.7.0] - 2019-06-06
 
 Thanks to @vlad-ro: